requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
azure-ai-formrecognizer==3.3.0
azure-identity==1.15.0
pydantic==2.5.0
aiofiles==23.2.1
jinja2==3.1.2 